Allow line drawing in float widths instead of just ints. Also means you can draw at <1 pixel wide [completed: 773]

+++ Shared/SCHPictureStarterDrawingInstruction.h
@@ -22,7 +22,7 @@
 @property (nonatomic, assign) NSInteger pointCount;
 @property (nonatomic, assign) NSInteger pointCapacity;
 @property (nonatomic, retain) UIColor *color;
-@property (nonatomic, assign) NSInteger size;
+@property (nonatomic, assign) CGFloat size;
 - (CGBlendMode)blendMode;
 @end
